Practica Done

+++ Framework/code/src/Shader.cpp
@@ -111,7 +111,7 @@ void Shader::SetBool(const std::string& name, bool value) const
 
 void Shader::SetInt(const std::string& name, int value) const
 {
-	glUniform1i(glGetUniformLocation(programID, name.c_str()), value);
+	glUniform1i(glGetUniformLocation(programID, name.c_str()), (int)value);
 }
 
 void Shader::SetFloat(const std::string& name, float value) const
